Separate base "panel" behaviour into parent class
